move EditArm6 to EditArm6Panel
